[dsch] new Result API
