+++ README.md
@@ -1,4 +1,7 @@
-# Result [![Coverage Status](https://coveralls.io/repos/github/DScheglov/ts-result/badge.svg?branch=main)](https://coveralls.io/github/DScheglov/ts-result?branch=main) [![npm version](https://img.shields.io/npm/v/@cardellini/ts-result.svg?style=flat-square)](https://www.npmjs.com/package/@cardellini/ts-result) [![npm downloads](https://img.shields.io/npm/dm/@cardellini/ts-result.svg?style=flat-square)](https://www.npmjs.com/package/@cardellini/ts-result) [![GitHub license](https://img.shields.io/badge/license-MIT-blue.svg)](https://github.com/DScheglov/ts-result/blob/master/LICENSE)
+# Result [![Coverage Status](https://coveralls.io/repos/github/DScheglov/ts-result/badge.svg?branch=main)](https://coveralls.io/github/DScheglov/ts-result?branch=main)
+[![npm version](https://img.shields.io/npm/v/@cardellini/ts-result.svg?style=flat-square)](https://www.npmjs.com/package/@cardellini/ts-result)
+[![npm downloads](https://img.shields.io/npm/dm/@cardellini/ts-result.svg?style=flat-square)](https://www.npmjs.com/package/@cardellini/ts-result)
+[![GitHub license](https://img.shields.io/badge/license-MIT-blue.svg)](https://github.com/DScheglov/ts-result/blob/master/LICENSE)
 
 Useful type to model success and failure, implemented with focus on type safety,
 developer experience and preserving flat learning curve.
@@ -109,10 +112,10 @@ console.log(silmarillionAuthors.unwrapErr());
 ### Working with Async Results
 
 ```typescript
-import { asyncDo, collect, err, ok } from '@cardellini/ts-result';
+import { Do, collect, err, ok } from '@cardellini/ts-result';
 
 const getBookWithAuthors = (bookId: string) =>
-  asyncDo(async function* () {
+  Do(async function* () {
     const book = yield* await fetchBook(bookId);
     const authors = yield* await fetchPersons(book.authorIds);
