+++ examples/book-and-authors.ts
@@ -1,16 +1,37 @@
-import { asyncDo, collect, err, ok } from '@cardellini/ts-result';
+import { AsyncResult, Do, collect, err, ok } from '@cardellini/ts-result';
+
+type Book = {
+  id: string;
+  title: string;
+  authorIds: string[];
+};
+
+type Person = {
+  id: string;
+  name: string;
+};
 
 const getBookWithAuthors = (bookId: string) =>
-  asyncDo(async function* () {
+  Do(async function* () {
     const book = yield* await fetchBook(bookId);
     const authors = yield* await fetchPersons(book.authorIds);
 
     return { ...book, authors };
   });
 
+const getBookWithAuthors2 = (bookId: string): AsyncResult<
+  Book & { authors: Person[] },
+  "ERR_BOOK_NOT_FOUND" | "ERR_PERSON_NOT_FOUND"
+> => Do(async function* () {
+  const book = yield* await fetchBook(bookId);
+  const authors = yield* await fetchPersons(book.authorIds);
+
+  return { ...book, authors };
+});
+
 const fetchBook = async (id: string) => (
-  id === '1' ? ok({ id, title: 'The Lord of the Rings', authorIds: ['1', '2'] }) :
-  id === '2' ? ok({ id, title: 'The Silmarillion', authorIds: ['1', '3'] }) :
+  id === '1' ? ok<Book>({ id, title: 'The Lord of the Rings', authorIds: ['1', '2'] }) :
+  id === '2' ? ok<Book>({ id, title: 'The Silmarillion', authorIds: ['1', '3'] }) :
   err('ERR_BOOK_NOT_FOUND' as const)
 );
