+++ src/async.ts
@@ -1,3 +1,3 @@
 export * from './base';
-export * from './async-do-gen';
 export * from './async-methods';
+export * from './do';
