+++ src/fn/is-promise.ts
@@ -1,0 +1,2 @@
+export const isPromise = (value: any): value is Promise<unknown> =>
+  value != null && typeof value.then === 'function';
