+++ src/index.test.ts
@@ -3,7 +3,6 @@ import * as index from './index';
 
 describe('index exports', () => {
   it.each([
-    index.asyncDo,
     index.sequenceAsync,
     index.thenMap,
     index.thenMapErr,
