+++ src/index.ts
@@ -1,6 +1,5 @@
 export * from './base';
-export * from './do-gen';
-export * from './async-do-gen';
+export * from './do';
 export * from './lists';
 export * from './sync-methods';
 export * from './async-methods';
