+++ src/sync.ts
@@ -1,4 +1,4 @@
 export * from './base';
 export * from './sync-methods';
-export * from './do-gen';
+export * from './do';
 export * from './lists';
