+++ src/types.test.ts
@@ -1,0 +1,8 @@
+import { describe, it, expect } from '@jest/globals';
+import * as Types from './types';
+
+describe('types', () => {
+  it('must be defined', () => {
+    expect(Types).toBeDefined();
+  });
+});
